# projectx/밤토리/fdd/openapi_v1.yaml
openapi: 3.0.3
info:
  title: VOICE AI Education Platform API
  version: "1.0.0"
  description: |
    API Specification for "Bamtori" (Voice AI Tutor for Kids).
    Based on FDD MVP v1.4 and Domain Ontology v0.1.1.
  contact:
    name: Backend Support Team
    email: support@bamtori.ai
servers:
  - url: https://api.bamtori.ai/v1
    description: Production Server
  - url: https://dev-api.bamtori.ai/v1
    description: Development Server
  - url: http://localhost:8000/v1
    description: Local Environment
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register new family account
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: {type: string, format: email}
                password: {type: string}
                parent_name: {type: string}
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  family_id: {type: string, format: uuid}
                  access_token: {type: string}
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: {type: string, format: email}
                password: {type: string}
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: {type: string}
                  family_id: {type: string, format: uuid}

  # --- Children ---
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []